+++ kubernetes-model-generator/kubernetes-model-core/src/main/java/io/fabric8/kubernetes/api/model/HasMetadata.java
@@ -358,6 +358,23 @@ public interface HasMetadata extends KubernetesResource {
           + getKind());
     }
 
+    if (!(owner instanceof GenericKubernetesResource)
+        && !(this instanceof GenericKubernetesResource)
+        && owner instanceof Namespaced) {
+      if (!(this instanceof Namespaced)) {
+        throw new IllegalArgumentException(
+            "Cannot add owner reference from a cluster scoped to a namespace scoped resource: "
+                + optionalMetadata().map(m -> "'" + m.getName() + "' ").orElse("unnamed ")
+                + getKind());
+      } else if (optionalMetadata().map(m -> !Objects.equals(m.getNamespace(), owner.getMetadata().getNamespace()))
+          .orElse(false)) {
+        throw new IllegalArgumentException(
+            "Cannot add owner reference between two resources in different namespaces:"
+                + optionalMetadata().map(m -> "'" + m.getName() + "' ").orElse("unnamed ")
+                + getKind());
+      }
+    }
+
     final OwnerReference ownerReference = new OwnerReferenceBuilder()
         .withUid(metadata.getUid())
         .withApiVersion(owner.getApiVersion())